# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exchange API",
        "description": "API documentation for ExchangeProject",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/owner-profits/export": {
            "get": {
                "tags": [
                    "Owner Profit"
                ],
                "summary": "Export Owner Profit records to Excel",
                "description": "Exports the list of Owner Profit records as an Excel file.",
                "operationId": "e8b264576b57769e543f31e8b24db6fb",
                "responses": {
                    "200": {
                        "description": "Excel file generated successfully."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/owner-profits": {
            "get": {
                "tags": [
                    "Owner Profit"
                ],
                "summary": "List Owner Profit records",
                "description": "Retrieve a paginated list of Owner Profit records for the admin view.",
                "operationId": "f9aeddea8db2f56c5d5811ba7f1d0f56",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            },
            "post": {
                "tags": [
                    "Owner Profit"
                ],
                "summary": "Add a new Owner Profit record",
                "description": "Create a new Owner Profit record.",
                "operationId": "eb28640dec1db68097892074c727d8cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cash_amount": {
                                        "description": "The amount of cash.",
                                        "type": "number"
                                    },
                                    "remarks": {
                                        "description": "Remarks for the record.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Owner Profit added successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction successfully added!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/owner-profits/{id}": {
            "delete": {
                "tags": [
                    "Owner Profit"
                ],
                "summary": "Delete an Owner Profit record",
                "description": "Delete a specific Owner Profit record by ID.",
                "operationId": "49c041a748b2a617f99cdc47ab2c3b64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Owner Profit deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Owner Profit deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Owner Profit not found."
                    }
                }
            }
        },
        "/exchange-owner-profits": {
            "get": {
                "tags": [
                    "Owner Profit"
                ],
                "summary": "List Owner Profit records for exchange",
                "description": "Retrieve a paginated list of Owner Profit records for the logged-in user's exchange.",
                "operationId": "2010c1f311a3e6655b3d9d15140547b2",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
...
    },
...
}